requirements.txt
# Core dependencies
pydantic-ai>=0.0.14
openai>=1.0.0
python-dotenv>=1.0.0

# Optional dependencies for advanced features
requests>=2.31.0
asyncio-mqtt>=0.16.0
schedule>=1.2.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging
colorlog>=6.7.0
